(2.6) pool queue plots - fix partial package refactoring and neglected removal of jndi arguments
Introduced by pull request https://github.com/dCache/dcache/pull/907.

See https://lists.dcache.org/sympa/rss/latest_arc/user-forum?count=20&for=10

The refactoring into new directories was not applied to the branch
when the patch was applied, and the build process did not catch
the incorrect package names which resulted.

Also neglected to eliminate two jndi properties, the presence of
which prevent httpd from booting.

Target: 2.6
Acked-by: Paul
Require-notes: yes
Require-book: no

RELEASE NOTES:

Fixes a bug in recent change to the pool queue plots which
prevents the httpd service from starting.
