+++ modules/webadmin/src/main/webapp/WEB-INF/WebAdminInterface.xml
@@ -36,8 +36,6 @@
     <jee:jndi-lookup id="alarmsCleanerSleepInterval" jndi-name="java:comp/env/alarmsCleanerSleepInterval"/>
     <jee:jndi-lookup id="alarmsCleanerDeleteThreshold" jndi-name="java:comp/env/alarmsCleanerDeleteThreshold"/>
     <jee:jndi-lookup id="pqpEnabled" jndi-name="java:comp/env/pqpEnabled"/>
-    <jee:jndi-lookup id="pqpRefreshInterval" jndi-name="java:comp/env/pqpRefreshInterval"/>
-    <jee:jndi-lookup id="pqpRefreshIntervalUnit" jndi-name="java:comp/env/pqpRefreshIntervalUnit"/>
     <jee:jndi-lookup id="rrdImgType" jndi-name="java:comp/env/rrdImgType"/>
     <jee:jndi-lookup id="rrdImgHeight" jndi-name="java:comp/env/rrdImgHeight"/>
     <jee:jndi-lookup id="rrdImgWidth" jndi-name="java:comp/env/rrdImgWidth"/>
@@ -125,7 +123,7 @@
         <property name="quitAtMaxDelay" value="false"/>
     </bean>
 
-    <bean id="RoundRobinDbSettings" class="org.dcache.webadmin.model.dataaccess.util.RrdSettings">
+    <bean id="RoundRobinDbSettings" class="org.dcache.webadmin.model.dataaccess.util.rrd4j.RrdSettings">
         <property name="baseDirectory" ref="plotsDir"/>
         <property name="imgType" ref="rrdImgType"/>
         <property name="imgHeight" ref="rrdImgHeight"/>
@@ -148,7 +146,7 @@
         <property name="heartbeatFactor" ref="rrdHeartbeatFactor"/>
     </bean>
 
-    <bean id="RrdPoolInfoAgent" class="org.dcache.webadmin.model.dataaccess.util.RrdPoolInfoAgent" init-method="initialize">
+    <bean id="RrdPoolInfoAgent" class="org.dcache.webadmin.model.dataaccess.util.rrd4j.RrdPoolInfoAgent" init-method="initialize">
         <property name="settings" ref="RoundRobinDbSettings"/>
     </bean>
